requirements.txt
fastapi
uvicorn
torch
torchvision
pillow
numpy
python-multipart
scikit-image
loguru
matplotlib